# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: GameReview API
  version: 1.0.0
  description: |
    Allows users to review and rate games that the admin adds, Anybody who plays video games, It would mean users could review and rate games on the client application.
    Every error uses RFC 9457 Problem Details (see components/schemas/Problem).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Team 03
    url: https://github.com/thomas-gonda/CST438Project2
servers:
  - url: http://localhost:8080
    description: Local (docker compose up)
  - url: https://your-api.example.com
    description: Production (replace in Sprint 3)

# Every endpoint requires a Bearer token unless it overrides this with `security: []`.
security:
  - bearerAuth: []
...
paths:
  /api/v1/games:
    get:
      tags: [Games]
      operationId: listGames
      summary: List available games with pagination
      description: |
        Public endpoint returning games ordered by ID ascending.
        Page-based pagination limits response size and lets clients request
        a specific page. Page numbers start at zero.
        An empty catalog or a page beyond the available results returns
        an empty content array with pagination metadata.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: One page of games with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamePage'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Games]
      operationId: createGame
      summary: Add a game to the catalog
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameCreate'
      responses:
        '201':
          description: Game created
          headers:
            Location:
              description: URL of the new game
              schema:
                type: string
                example: /api/v1/games/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from the OAuth2 Authorization Code + PKCE flow (with Google, send the ID token; Google's access_token is opaque).
    basicAuth:
      type: http
      scheme: basic
      description: Local admin authentication used by the game creation endpoint until OAuth is connected.
  parameters:
...
    Page:
      name: page
      in: query
      description: Zero-based page number
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    Game:
      type: object
      required: [id, name, publisher, category, price]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          minLength: 1
          example: Stardew Valley
        publisher:
          type: string
          minLength: 1
          example: ConcernedApe
        category:
          type: string
          minLength: 1
          example: Simulation
        price:
          type: number
          minimum: 0
          example: 14.99
    GameCreate:
      type: object
      required: [name, publisher, category, price]
      properties:
        name:
          type: string
          minLength: 1
        publisher:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        price:
          type: number
          format: double
          minimum: 0
...
    PageMeta:
      description: Spring Data PagedModel metadata.
      type: object
      required: [size, number, totalElements, totalPages]
      properties:
        size:
          type: integer
          example: 20
        number:
          type: integer
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 57
        totalPages:
          type: integer
          example: 3
...
    GamePage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        page:
          $ref: '#/components/schemas/PageMeta'
...
    Problem:
      description: RFC 9457 Problem Details. The common error body returned by the API.
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          default: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Game 42 not found
        instance:
          type: string
          example: /api/v1/games/42
  responses:
    BadRequest:
      description: Validation failed or malformed JSON
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Bad Request
            status: 400
            detail: One or more fields failed validation. Check the request against the contract.
    Unauthorized:
      description: Missing or invalid token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Unauthorized
            status: 401
    Forbidden:
      description: Valid token, but the caller is not allowed to perform this operation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Forbidden
            status: 403
...